Move files in fic directory to utils directory
