+++ fic/utils/auth_env.go
@@ -1,4 +1,4 @@
-package fic
+package utils
 
 import (
 	"os"
