+++ fic/utils/endpoint_location.go
@@ -1,4 +1,4 @@
-package fic
+package utils
 
 import (
 	fic "github.com/nttcom/go-fic"
